Migrating actually to CCR\AppKernel;
