+++ classes/CCR/AppKernel/CurrentQueue.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace AppKernel;
+namespace CCR\AppKernel;
 
 use CCR\DB;
