+++ classes/CCR/AppKernel/InstanceData.php
@@ -5,7 +5,7 @@
 // parameters.
 // ================================================================================
 
-namespace AppKernel;
+namespace CCR\AppKernel;
 
 // ================================================================================
 // Application kernel data.  Data is classified into three categories and named
@@ -363,223 +363,5 @@ class InstanceData
 
         $str = $unzippedStr;
         return true;
-    }
-}
-
-/**
- * Class InstanceMetric Application kernel metric/statistics
- * @package AppKernel
- */
-class InstanceMetric
-{
-    /**
-     * @var int|null Database id of the metric.
-     */
-    public $id = null;
-
-    // Metric data parsed from the app kernel
-    /**
-     * @var string|null Metrics name
-     */
-    public $name = null;
-
-    /**
-     * @var string|null  Metrics value
-     */
-    public $value = null;
-
-    /**
-     * @var string|null  Metrics units
-     */
-    public $unit = null;
-
-    // --------------------------------------------------------------------------------
-
-    public static function sort_cmp(InstanceMetric $a, InstanceMetric $b)
-    {
-        return strcmp($a->name, $b->name);
-    }
-
-    // --------------------------------------------------------------------------------
-
-    public function __construct($name, $value, $unit = null, $id = null)
-    {
-        if($id!==null) {
-            $this->id = intval($id);
-        }
-        $this->name = $name;
-        $this->value = $value;
-        $this->unit = $unit;
-    }
-
-    // --------------------------------------------------------------------------------
-
-    public function __toString()
-    {
-        return __CLASS__ . ": ({$this->name}, {$this->value}, {$this->unit})";
-    }
-
-    // --------------------------------------------------------------------------------
-    // Generate the metric guid to uniquely identify this metric.
-    //
-    // @returns A unique identifier for the metric
-    // --------------------------------------------------------------------------------
-
-    public function guid()
-    {
-        return md5($this->name . $this->unit);
-    }
-}  // class InstanceMetric
-
-// ================================================================================
-// Application knernel parameters (currently the same info as the metrics).
-// ================================================================================
-
-class InstanceParameter extends InstanceMetric
-{
-    // Optional grouping tag supplied with the parameter (e.g., App in App:ExeBinSignature)
-    public $tag = null;
-
-    // --------------------------------------------------------------------------------
-
-    public function __construct($name, $value, $unit = null, $tag = null)
-    {
-
-        // If the tag was explicitly provided us it, otherwise check to see if it is
-        // still in the name.
-
-        if (null !== $tag) {
-            $this->name = $name;
-            $this->tag = $tag;
-        } elseif (false !== strpos($name, ":")) {
-            $parts = explode(":", $name);
-            $this->tag = array_shift($parts);
-            $this->name = implode(":", $parts);
-        } else {
-            $this->name = $name;
-        }
-
-        InstanceData::decode($value);
-        $this->value = $value;
-        $this->unit = $unit;
-    }
-
-    // --------------------------------------------------------------------------------
-    // Generate the metric guid to uniquely identify this metric.
-    //
-    // @returns A unique identifier for the metric
-    // --------------------------------------------------------------------------------
-
-    public function guid()
-    {
-        return md5((null === $this->tag ? "" : $this->tag) . $this->name . $this->unit);
-    }
-}
-
-
-/**
- * Class AppKernelDefinition Application kernel definition
- * @package AppKernel
- */
-class AppKernelDefinition
-{
-    /**
-     * @var int|null Database id
-     */
-    public $id = null;
-
-    /**
-     * @var string|null Public name
-     */
-    public $name = null;
-
-    /**
-     * @var string|null  Reporter basename
-     */
-    public $basename = null;
-
-    /**
-     * @var string|null App kernel description (html)
-     */
-    public $description = null;
-
-    /**
-     * @var string|null Processor_units node or core
-     */
-    public $processor_unit = null;
-
-    /**
-     * @var bool true if enabled
-     */
-    public $enabled = false;
-
-    /**
-     * @var bool  true if visible outside of the application kernel tools
-     */
-    public $visible = false;
-
-    /**
-     * @var int|null start time (can be limited to conditions like resource)
-     */
-    public $start_ts = null;
-
-    /**
-     * @var int|null end time (can be limited to conditions like resource)
-     */
-    public $end_ts = null;
-
-    public function __construct(
-        $id = null,
-        $name = null,
-        $basename = null,
-        $description = null,
-        $processor_unit = null,
-        $enabled = false,
-        $visible = false,
-        $start_ts = null,
-        $end_ts = null
-    ) {
-        $this->id = intval($id);
-        $this->name = $name;
-        $this->basename = $basename;
-        $this->description = $description;
-        $this->processor_unit = $processor_unit;
-        $this->enabled = (1 == $enabled ? true : false);
-        $this->visible = (1 == $visible ? true : false);
-        $this->start_ts = intval($start_ts);
-        $this->end_ts = intval($end_ts);
-    }
-}
-
-// ================================================================================
-// Resource that application kernels are run on
-// ================================================================================
-
-class AKResource
-{
-    public $id = null;  // Database id
-    public $nickname = null;  // Resource nickname (e.g., short name such as "edge")
-    public $name = null;  // Resource full name (e.g., edge.ccr.buffalo.edu)
-    public $description = null;  // Resource description
-    public $enabled = false;  // true if the resoure is enabled
-    public $visible = false;  // true if visible outside of the application kernel tools
-    public $xdmod_resource_id = null;  // XDMoD resource id
-}
-
-class ProcessingUnit
-{
-    /**
-     * @var string|null
-     */
-    public $unit = null;
-    /**
-     * @var int|null
-     */
-    public $count = null;
-
-    public function __construct($unit = null, $count = null)
-    {
-        $this->unit = $unit;
-        $this->count = intval($count);
     }
 }
