+++ classes/CCR/AppKernel/PerformanceMap.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace AppKernel;
+namespace CCR\AppKernel;
 
 use CCR\DB;
-use xd_utilities;
-use DateTime;
 use DateInterval;
+use DateTime;
+use xd_utilities;
 
 class PerformanceMap
 {
