+++ classes/CCR/AppKernel/ProblemDetector/ResourceLevel.php
