+++ classes/CCR/AppKernel/Query.php
@@ -1,5 +1,5 @@
 <?php
-namespace AppKernel;
+namespace CCR\AppKernel;
 
 use CCR\DB;
