+++ classes/CCR/AppKernel/RunsStats.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace AppKernel;
+namespace CCR\AppKernel;
 
 class RunsStats
 {
