+++ src/Controller/AkrrController.php
@@ -2,6 +2,7 @@
 
 namespace CCR\Controller;
 
+use HttpInvalidParamException;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
